- UI theme fixes and improvements - API code change improvements - Updated search bar of the app - UX improvements
