+++ app/src/main/java/com/th3pl4gu3/mauritius_emergency_services/ui/theme/Color.kt
@@ -1,56 +1,61 @@
 package com.th3pl4gu3.mauritius_emergency_services.ui.theme
 
+import androidx.compose.material3.darkColorScheme
+import androidx.compose.material3.lightColorScheme
+import androidx.compose.runtime.Immutable
 import androidx.compose.ui.graphics.Color
 
-val EmergencyButton = Color(0xFFEF5350)
+/**
+ * Color definitions
+ **/
 
 val primaryLight = Color(0xFF3B608F)
 val onPrimaryLight = Color(0xFFFFFFFF)
 val primaryContainerLight = Color(0xFFD4E3FF)
 val onPrimaryContainerLight = Color(0xFF001C39)
-val secondaryLight = Color(0xFF5C6380)
+val secondaryLight = Color(0xFF545F71)
 val onSecondaryLight = Color(0xFFFFFFFF)
-val secondaryContainerLight = Color(0xFFD4E3FF)
-val onSecondaryContainerLight = Color(0xFF001C3A)
-val tertiaryLight = Color(0xFF3A608F)
+val secondaryContainerLight = Color(0xFFD8E3F8)
+val onSecondaryContainerLight = Color(0xFF111C2B)
+val tertiaryLight = Color(0xFF7F4D7A)
 val onTertiaryLight = Color(0xFFFFFFFF)
-val tertiaryContainerLight = Color(0xFFD3E3FF)
-val onTertiaryContainerLight = Color(0xFF001C39)
-val errorLight = Color(0xFFEF5350)
+val tertiaryContainerLight = Color(0xFFFFD7F6)
+val onTertiaryContainerLight = Color(0xFF330833)
+val errorLight = Color(0xFFBA1A1A)
 val onErrorLight = Color(0xFFFFFFFF)
 val errorContainerLight = Color(0xFFFFDAD6)
 val onErrorContainerLight = Color(0xFF410002)
 val backgroundLight = Color(0xFFF8F9FF)
 val onBackgroundLight = Color(0xFF191C20)
-val surfaceLight = Color(0xFFF8F9FF)
-val onSurfaceLight = Color(0xFF191C20)
-val surfaceVariantLight = Color(0xFFE1E2EC)
-val onSurfaceVariantLight = Color(0xFF44474F)
-val outlineLight = Color(0xFF75777F)
-val outlineVariantLight = Color(0xFFC4C6D0)
+val surfaceLight = Color(0xFFFAF8FF)
+val onSurfaceLight = Color(0xFF1A1B21)
+val surfaceVariantLight = Color(0xFFDDE3EA)
+val onSurfaceVariantLight = Color(0xFF41484D)
+val outlineLight = Color(0xFF71787E)
+val outlineVariantLight = Color(0xFFC1C7CD)
 val scrimLight = Color(0xFF000000)
-val inverseSurfaceLight = Color(0xFF2E3035)
-val inverseOnSurfaceLight = Color(0xFFEFF0F7)
+val inverseSurfaceLight = Color(0xFF2F3036)
+val inverseOnSurfaceLight = Color(0xFFF1F0F7)
 val inversePrimaryLight = Color(0xFFA4C9FE)
-val surfaceDimLight = Color(0xFFD8DAE0)
-val surfaceBrightLight = Color(0xFFF8F9FF)
+val surfaceDimLight = Color(0xFFDAD9E0)
+val surfaceBrightLight = Color(0xFFFAF8FF)
 val surfaceContainerLowestLight = Color(0xFFFFFFFF)
-val surfaceContainerLowLight = Color(0xFFF2F3FA)
-val surfaceContainerLight = Color(0xFFEDEDF4)
-val surfaceContainerHighLight = Color(0xFFE7E8EE)
-val surfaceContainerHighestLight = Color(0xFFE1E2E8)
+val surfaceContainerLowLight = Color(0xFFF4F3FA)
+val surfaceContainerLight = Color(0xFFEEEDF4)
+val surfaceContainerHighLight = Color(0xFFE8E7EF)
+val surfaceContainerHighestLight = Color(0xFFE2E2E9)
 
 val primaryLightMediumContrast = Color(0xFF1B4472)
 val onPrimaryLightMediumContrast = Color(0xFFFFFFFF)
 val primaryContainerLightMediumContrast = Color(0xFF5276A7)
 val onPrimaryContainerLightMediumContrast = Color(0xFFFFFFFF)
-val secondaryLightMediumContrast = Color(0xFF1C4472)
+val secondaryLightMediumContrast = Color(0xFF394354)
 val onSecondaryLightMediumContrast = Color(0xFFFFFFFF)
-val secondaryContainerLightMediumContrast = Color(0xFF5276A7)
+val secondaryContainerLightMediumContrast = Color(0xFF6A7587)
 val onSecondaryContainerLightMediumContrast = Color(0xFFFFFFFF)
-val tertiaryLightMediumContrast = Color(0xFF1A4472)
+val tertiaryLightMediumContrast = Color(0xFF61325D)
 val onTertiaryLightMediumContrast = Color(0xFFFFFFFF)
-val tertiaryContainerLightMediumContrast = Color(0xFF5176A7)
+val tertiaryContainerLightMediumContrast = Color(0xFF986391)
 val onTertiaryContainerLightMediumContrast = Color(0xFFFFFFFF)
 val errorLightMediumContrast = Color(0xFF8C0009)
 val onErrorLightMediumContrast = Color(0xFFFFFFFF)
@@ -58,35 +63,35 @@ val errorContainerLightMediumContrast = Color(0xFFDA342E)
 val onErrorContainerLightMediumContrast = Color(0xFFFFFFFF)
 val backgroundLightMediumContrast = Color(0xFFF8F9FF)
 val onBackgroundLightMediumContrast = Color(0xFF191C20)
-val surfaceLightMediumContrast = Color(0xFFF8F9FF)
-val onSurfaceLightMediumContrast = Color(0xFF191C20)
-val surfaceVariantLightMediumContrast = Color(0xFFE1E2EC)
-val onSurfaceVariantLightMediumContrast = Color(0xFF40434B)
-val outlineLightMediumContrast = Color(0xFF5C5F67)
-val outlineVariantLightMediumContrast = Color(0xFF787A83)
+val surfaceLightMediumContrast = Color(0xFFFAF8FF)
+val onSurfaceLightMediumContrast = Color(0xFF1A1B21)
+val surfaceVariantLightMediumContrast = Color(0xFFDDE3EA)
+val onSurfaceVariantLightMediumContrast = Color(0xFF3D4449)
+val outlineLightMediumContrast = Color(0xFF596065)
+val outlineVariantLightMediumContrast = Color(0xFF757C81)
 val scrimLightMediumContrast = Color(0xFF000000)
-val inverseSurfaceLightMediumContrast = Color(0xFF2E3035)
-val inverseOnSurfaceLightMediumContrast = Color(0xFFEFF0F7)
+val inverseSurfaceLightMediumContrast = Color(0xFF2F3036)
+val inverseOnSurfaceLightMediumContrast = Color(0xFFF1F0F7)
 val inversePrimaryLightMediumContrast = Color(0xFFA4C9FE)
-val surfaceDimLightMediumContrast = Color(0xFFD8DAE0)
-val surfaceBrightLightMediumContrast = Color(0xFFF8F9FF)
+val surfaceDimLightMediumContrast = Color(0xFFDAD9E0)
+val surfaceBrightLightMediumContrast = Color(0xFFFAF8FF)
 val surfaceContainerLowestLightMediumContrast = Color(0xFFFFFFFF)
-val surfaceContainerLowLightMediumContrast = Color(0xFFF2F3FA)
-val surfaceContainerLightMediumContrast = Color(0xFFEDEDF4)
-val surfaceContainerHighLightMediumContrast = Color(0xFFE7E8EE)
-val surfaceContainerHighestLightMediumContrast = Color(0xFFE1E2E8)
+val surfaceContainerLowLightMediumContrast = Color(0xFFF4F3FA)
+val surfaceContainerLightMediumContrast = Color(0xFFEEEDF4)
+val surfaceContainerHighLightMediumContrast = Color(0xFFE8E7EF)
+val surfaceContainerHighestLightMediumContrast = Color(0xFFE2E2E9)
 
-val primaryLightHighContrast = Color(0xFF002345)
+val primaryLightHighContrast = Color(0xFF002344)
 val onPrimaryLightHighContrast = Color(0xFFFFFFFF)
 val primaryContainerLightHighContrast = Color(0xFF1B4472)
 val onPrimaryContainerLightHighContrast = Color(0xFFFFFFFF)
-val secondaryLightHighContrast = Color(0xFF002345)
+val secondaryLightHighContrast = Color(0xFF182332)
 val onSecondaryLightHighContrast = Color(0xFFFFFFFF)
-val secondaryContainerLightHighContrast = Color(0xFF1C4472)
+val secondaryContainerLightHighContrast = Color(0xFF394354)
 val onSecondaryContainerLightHighContrast = Color(0xFFFFFFFF)
-val tertiaryLightHighContrast = Color(0xFF002344)
+val tertiaryLightHighContrast = Color(0xFF3B103A)
 val onTertiaryLightHighContrast = Color(0xFFFFFFFF)
-val tertiaryContainerLightHighContrast = Color(0xFF1A4472)
+val tertiaryContainerLightHighContrast = Color(0xFF61325D)
 val onTertiaryContainerLightHighContrast = Color(0xFFFFFFFF)
 val errorLightHighContrast = Color(0xFF4E0002)
 val onErrorLightHighContrast = Color(0xFFFFFFFF)
@@ -94,71 +99,71 @@ val errorContainerLightHighContrast = Color(0xFF8C0009)
 val onErrorContainerLightHighContrast = Color(0xFFFFFFFF)
 val backgroundLightHighContrast = Color(0xFFF8F9FF)
 val onBackgroundLightHighContrast = Color(0xFF191C20)
-val surfaceLightHighContrast = Color(0xFFF8F9FF)
+val surfaceLightHighContrast = Color(0xFFFAF8FF)
 val onSurfaceLightHighContrast = Color(0xFF000000)
-val surfaceVariantLightHighContrast = Color(0xFFE1E2EC)
-val onSurfaceVariantLightHighContrast = Color(0xFF21242B)
-val outlineLightHighContrast = Color(0xFF40434B)
-val outlineVariantLightHighContrast = Color(0xFF40434B)
+val surfaceVariantLightHighContrast = Color(0xFFDDE3EA)
+val onSurfaceVariantLightHighContrast = Color(0xFF1E2529)
+val outlineLightHighContrast = Color(0xFF3D4449)
+val outlineVariantLightHighContrast = Color(0xFF3D4449)
 val scrimLightHighContrast = Color(0xFF000000)
-val inverseSurfaceLightHighContrast = Color(0xFF2E3035)
+val inverseSurfaceLightHighContrast = Color(0xFF2F3036)
 val inverseOnSurfaceLightHighContrast = Color(0xFFFFFFFF)
 val inversePrimaryLightHighContrast = Color(0xFFE3ECFF)
-val surfaceDimLightHighContrast = Color(0xFFD8DAE0)
-val surfaceBrightLightHighContrast = Color(0xFFF8F9FF)
+val surfaceDimLightHighContrast = Color(0xFFDAD9E0)
+val surfaceBrightLightHighContrast = Color(0xFFFAF8FF)
 val surfaceContainerLowestLightHighContrast = Color(0xFFFFFFFF)
-val surfaceContainerLowLightHighContrast = Color(0xFFF2F3FA)
-val surfaceContainerLightHighContrast = Color(0xFFEDEDF4)
-val surfaceContainerHighLightHighContrast = Color(0xFFE7E8EE)
-val surfaceContainerHighestLightHighContrast = Color(0xFFE1E2E8)
+val surfaceContainerLowLightHighContrast = Color(0xFFF4F3FA)
+val surfaceContainerLightHighContrast = Color(0xFFEEEDF4)
+val surfaceContainerHighLightHighContrast = Color(0xFFE8E7EF)
+val surfaceContainerHighestLightHighContrast = Color(0xFFE2E2E9)
 
 val primaryDark = Color(0xFFA4C9FE)
 val onPrimaryDark = Color(0xFF00315D)
 val primaryContainerDark = Color(0xFF204876)
 val onPrimaryContainerDark = Color(0xFFD4E3FF)
-val secondaryDark = Color(0xFFA5C8FE)
-val onSecondaryDark = Color(0xFF00315D)
-val secondaryContainerDark = Color(0xFF214876)
-val onSecondaryContainerDark = Color(0xFFD4E3FF)
-val tertiaryDark = Color(0xFFA4C9FE)
-val onTertiaryDark = Color(0xFF00315C)
-val tertiaryContainerDark = Color(0xFF1F4876)
-val onTertiaryContainerDark = Color(0xFFD3E3FF)
+val secondaryDark = Color(0xFFBCC7DB)
+val onSecondaryDark = Color(0xFF263141)
+val secondaryContainerDark = Color(0xFF3C4758)
+val onSecondaryContainerDark = Color(0xFFD8E3F8)
+val tertiaryDark = Color(0xFFF1B3E7)
+val onTertiaryDark = Color(0xFF4C1F49)
+val tertiaryContainerDark = Color(0xFF653661)
+val onTertiaryContainerDark = Color(0xFFFFD7F6)
 val errorDark = Color(0xFFFFB4AB)
 val onErrorDark = Color(0xFF690005)
 val errorContainerDark = Color(0xFF93000A)
 val onErrorContainerDark = Color(0xFFFFDAD6)
 val backgroundDark = Color(0xFF111318)
 val onBackgroundDark = Color(0xFFE1E2E9)
-val surfaceDark = Color(0xFF111318)
-val onSurfaceDark = Color(0xFFE1E2E8)
-val surfaceVariantDark = Color(0xFF44474F)
-val onSurfaceVariantDark = Color(0xFFC4C6D0)
-val outlineDark = Color(0xFF8E9099)
-val outlineVariantDark = Color(0xFF44474F)
+val surfaceDark = Color(0xFF121318)
+val onSurfaceDark = Color(0xFFE2E2E9)
+val surfaceVariantDark = Color(0xFF41484D)
+val onSurfaceVariantDark = Color(0xFFC1C7CD)
+val outlineDark = Color(0xFF8B9297)
+val outlineVariantDark = Color(0xFF41484D)
 val scrimDark = Color(0xFF000000)
-val inverseSurfaceDark = Color(0xFFE1E2E8)
-val inverseOnSurfaceDark = Color(0xFF2E3035)
+val inverseSurfaceDark = Color(0xFFE2E2E9)
+val inverseOnSurfaceDark = Color(0xFF2F3036)
 val inversePrimaryDark = Color(0xFF3B608F)
-val surfaceDimDark = Color(0xFF111318)
-val surfaceBrightDark = Color(0xFF37393E)
-val surfaceContainerLowestDark = Color(0xFF0C0E13)
-val surfaceContainerLowDark = Color(0xFF191C20)
-val surfaceContainerDark = Color(0xFF1D2024)
-val surfaceContainerHighDark = Color(0xFF272A2F)
-val surfaceContainerHighestDark = Color(0xFF32353A)
+val surfaceDimDark = Color(0xFF121318)
+val surfaceBrightDark = Color(0xFF38393F)
+val surfaceContainerLowestDark = Color(0xFF0D0E13)
+val surfaceContainerLowDark = Color(0xFF1A1B21)
+val surfaceContainerDark = Color(0xFF1E1F25)
+val surfaceContainerHighDark = Color(0xFF282A2F)
+val surfaceContainerHighestDark = Color(0xFF33343A)
 
 val primaryDarkMediumContrast = Color(0xFFABCDFF)
 val onPrimaryDarkMediumContrast = Color(0xFF001730)
-val primaryContainerDarkMediumContrast = Color(0xFF6F93C5)
+val primaryContainerDarkMediumContrast = Color(0xFF6E93C5)
 val onPrimaryContainerDarkMediumContrast = Color(0xFF000000)
-val secondaryDarkMediumContrast = Color(0xFFACCDFF)
-val onSecondaryDarkMediumContrast = Color(0xFF001730)
-val secondaryContainerDarkMediumContrast = Color(0xFF6F92C5)
+val secondaryDarkMediumContrast = Color(0xFFC0CBE0)
+val onSecondaryDarkMediumContrast = Color(0xFF0B1726)
+val secondaryContainerDarkMediumContrast = Color(0xFF8691A4)
 val onSecondaryContainerDarkMediumContrast = Color(0xFF000000)
-val tertiaryDarkMediumContrast = Color(0xFFABCDFF)
-val onTertiaryDarkMediumContrast = Color(0xFF001730)
-val tertiaryContainerDarkMediumContrast = Color(0xFF6E93C5)
+val tertiaryDarkMediumContrast = Color(0xFFF5B7EB)
+val onTertiaryDarkMediumContrast = Color(0xFF2D032D)
+val tertiaryContainerDarkMediumContrast = Color(0xFFB67EAF)
 val onTertiaryContainerDarkMediumContrast = Color(0xFF000000)
 val errorDarkMediumContrast = Color(0xFFFFBAB1)
 val onErrorDarkMediumContrast = Color(0xFF370001)
@@ -166,23 +171,23 @@ val errorContainerDarkMediumContrast = Color(0xFFFF5449)
 val onErrorContainerDarkMediumContrast = Color(0xFF000000)
 val backgroundDarkMediumContrast = Color(0xFF111318)
 val onBackgroundDarkMediumContrast = Color(0xFFE1E2E9)
-val surfaceDarkMediumContrast = Color(0xFF111318)
-val onSurfaceDarkMediumContrast = Color(0xFFFBFAFF)
-val surfaceVariantDarkMediumContrast = Color(0xFF44474F)
-val onSurfaceVariantDarkMediumContrast = Color(0xFFC9CAD4)
-val outlineDarkMediumContrast = Color(0xFFA1A2AC)
-val outlineVariantDarkMediumContrast = Color(0xFF81838C)
+val surfaceDarkMediumContrast = Color(0xFF121318)
+val onSurfaceDarkMediumContrast = Color(0xFFFCFAFF)
+val surfaceVariantDarkMediumContrast = Color(0xFF41484D)
+val onSurfaceVariantDarkMediumContrast = Color(0xFFC5CCD2)
+val outlineDarkMediumContrast = Color(0xFF9DA4AA)
+val outlineVariantDarkMediumContrast = Color(0xFF7D848A)
 val scrimDarkMediumContrast = Color(0xFF000000)
-val inverseSurfaceDarkMediumContrast = Color(0xFFE1E2E8)
-val inverseOnSurfaceDarkMediumContrast = Color(0xFF272A2F)
-val inversePrimaryDarkMediumContrast = Color(0xFF224977)
-val surfaceDimDarkMediumContrast = Color(0xFF111318)
-val surfaceBrightDarkMediumContrast = Color(0xFF37393E)
-val surfaceContainerLowestDarkMediumContrast = Color(0xFF0C0E13)
-val surfaceContainerLowDarkMediumContrast = Color(0xFF191C20)
-val surfaceContainerDarkMediumContrast = Color(0xFF1D2024)
-val surfaceContainerHighDarkMediumContrast = Color(0xFF272A2F)
-val surfaceContainerHighestDarkMediumContrast = Color(0xFF32353A)
+val inverseSurfaceDarkMediumContrast = Color(0xFFE2E2E9)
+val inverseOnSurfaceDarkMediumContrast = Color(0xFF282A2F)
+val inversePrimaryDarkMediumContrast = Color(0xFF214977)
+val surfaceDimDarkMediumContrast = Color(0xFF121318)
+val surfaceBrightDarkMediumContrast = Color(0xFF38393F)
+val surfaceContainerLowestDarkMediumContrast = Color(0xFF0D0E13)
+val surfaceContainerLowDarkMediumContrast = Color(0xFF1A1B21)
+val surfaceContainerDarkMediumContrast = Color(0xFF1E1F25)
+val surfaceContainerHighDarkMediumContrast = Color(0xFF282A2F)
+val surfaceContainerHighestDarkMediumContrast = Color(0xFF33343A)
 
 val primaryDarkHighContrast = Color(0xFFFBFAFF)
 val onPrimaryDarkHighContrast = Color(0xFF000000)
@@ -190,11 +195,11 @@ val primaryContainerDarkHighContrast = Color(0xFFABCDFF)
 val onPrimaryContainerDarkHighContrast = Color(0xFF000000)
 val secondaryDarkHighContrast = Color(0xFFFBFAFF)
 val onSecondaryDarkHighContrast = Color(0xFF000000)
-val secondaryContainerDarkHighContrast = Color(0xFFACCDFF)
+val secondaryContainerDarkHighContrast = Color(0xFFC0CBE0)
 val onSecondaryContainerDarkHighContrast = Color(0xFF000000)
-val tertiaryDarkHighContrast = Color(0xFFFBFAFF)
+val tertiaryDarkHighContrast = Color(0xFFFFF9FA)
 val onTertiaryDarkHighContrast = Color(0xFF000000)
-val tertiaryContainerDarkHighContrast = Color(0xFFABCDFF)
+val tertiaryContainerDarkHighContrast = Color(0xFFF5B7EB)
 val onTertiaryContainerDarkHighContrast = Color(0xFF000000)
 val errorDarkHighContrast = Color(0xFFFFF9F9)
 val onErrorDarkHighContrast = Color(0xFF000000)
@@ -202,20 +207,251 @@ val errorContainerDarkHighContrast = Color(0xFFFFBAB1)
 val onErrorContainerDarkHighContrast = Color(0xFF000000)
 val backgroundDarkHighContrast = Color(0xFF111318)
 val onBackgroundDarkHighContrast = Color(0xFFE1E2E9)
-val surfaceDarkHighContrast = Color(0xFF111318)
+val surfaceDarkHighContrast = Color(0xFF121318)
 val onSurfaceDarkHighContrast = Color(0xFFFFFFFF)
-val surfaceVariantDarkHighContrast = Color(0xFF44474F)
-val onSurfaceVariantDarkHighContrast = Color(0xFFFBFAFF)
-val outlineDarkHighContrast = Color(0xFFC9CAD4)
-val outlineVariantDarkHighContrast = Color(0xFFC9CAD4)
+val surfaceVariantDarkHighContrast = Color(0xFF41484D)
+val onSurfaceVariantDarkHighContrast = Color(0xFFF8FBFF)
+val outlineDarkHighContrast = Color(0xFFC5CCD2)
+val outlineVariantDarkHighContrast = Color(0xFFC5CCD2)
 val scrimDarkHighContrast = Color(0xFF000000)
-val inverseSurfaceDarkHighContrast = Color(0xFFE1E2E8)
+val inverseSurfaceDarkHighContrast = Color(0xFFE2E2E9)
 val inverseOnSurfaceDarkHighContrast = Color(0xFF000000)
 val inversePrimaryDarkHighContrast = Color(0xFF002B52)
-val surfaceDimDarkHighContrast = Color(0xFF111318)
-val surfaceBrightDarkHighContrast = Color(0xFF37393E)
-val surfaceContainerLowestDarkHighContrast = Color(0xFF0C0E13)
-val surfaceContainerLowDarkHighContrast = Color(0xFF191C20)
-val surfaceContainerDarkHighContrast = Color(0xFF1D2024)
-val surfaceContainerHighDarkHighContrast = Color(0xFF272A2F)
-val surfaceContainerHighestDarkHighContrast = Color(0xFF32353A)
+val surfaceDimDarkHighContrast = Color(0xFF121318)
+val surfaceBrightDarkHighContrast = Color(0xFF38393F)
+val surfaceContainerLowestDarkHighContrast = Color(0xFF0D0E13)
+val surfaceContainerLowDarkHighContrast = Color(0xFF1A1B21)
+val surfaceContainerDarkHighContrast = Color(0xFF1E1F25)
+val surfaceContainerHighDarkHighContrast = Color(0xFF282A2F)
+val surfaceContainerHighestDarkHighContrast = Color(0xFF33343A)
+
+/**
+ * Color schemes
+ **/
+internal val lightScheme = lightColorScheme(
+    primary = primaryLight,
+    onPrimary = onPrimaryLight,
+    primaryContainer = primaryContainerLight,
+    onPrimaryContainer = onPrimaryContainerLight,
+    secondary = secondaryLight,
+    onSecondary = onSecondaryLight,
+    secondaryContainer = secondaryContainerLight,
+    onSecondaryContainer = onSecondaryContainerLight,
+    tertiary = tertiaryLight,
+    onTertiary = onTertiaryLight,
+    tertiaryContainer = tertiaryContainerLight,
+    onTertiaryContainer = onTertiaryContainerLight,
+    error = errorLight,
+    onError = onErrorLight,
+    errorContainer = errorContainerLight,
+    onErrorContainer = onErrorContainerLight,
+    background = backgroundLight,
+    onBackground = onBackgroundLight,
+    surface = surfaceLight,
+    onSurface = onSurfaceLight,
+    surfaceVariant = surfaceVariantLight,
+    onSurfaceVariant = onSurfaceVariantLight,
+    outline = outlineLight,
+    outlineVariant = outlineVariantLight,
+    scrim = scrimLight,
+    inverseSurface = inverseSurfaceLight,
+    inverseOnSurface = inverseOnSurfaceLight,
+    inversePrimary = inversePrimaryLight,
+    surfaceDim = surfaceDimLight,
+    surfaceBright = surfaceBrightLight,
+    surfaceContainerLowest = surfaceContainerLowestLight,
+    surfaceContainerLow = surfaceContainerLowLight,
+    surfaceContainer = surfaceContainerLight,
+    surfaceContainerHigh = surfaceContainerHighLight,
+    surfaceContainerHighest = surfaceContainerHighestLight,
+)
+
+internal val darkScheme = darkColorScheme(
+    primary = primaryDark,
+    onPrimary = onPrimaryDark,
+    primaryContainer = primaryContainerDark,
+    onPrimaryContainer = onPrimaryContainerDark,
+    secondary = secondaryDark,
+    onSecondary = onSecondaryDark,
+    secondaryContainer = secondaryContainerDark,
+    onSecondaryContainer = onSecondaryContainerDark,
+    tertiary = tertiaryDark,
+    onTertiary = onTertiaryDark,
+    tertiaryContainer = tertiaryContainerDark,
+    onTertiaryContainer = onTertiaryContainerDark,
+    error = errorDark,
+    onError = onErrorDark,
+    errorContainer = errorContainerDark,
+    onErrorContainer = onErrorContainerDark,
+    background = backgroundDark,
+    onBackground = onBackgroundDark,
+    surface = surfaceDark,
+    onSurface = onSurfaceDark,
+    surfaceVariant = surfaceVariantDark,
+    onSurfaceVariant = onSurfaceVariantDark,
+    outline = outlineDark,
+    outlineVariant = outlineVariantDark,
+    scrim = scrimDark,
+    inverseSurface = inverseSurfaceDark,
+    inverseOnSurface = inverseOnSurfaceDark,
+    inversePrimary = inversePrimaryDark,
+    surfaceDim = surfaceDimDark,
+    surfaceBright = surfaceBrightDark,
+    surfaceContainerLowest = surfaceContainerLowestDark,
+    surfaceContainerLow = surfaceContainerLowDark,
+    surfaceContainer = surfaceContainerDark,
+    surfaceContainerHigh = surfaceContainerHighDark,
+    surfaceContainerHighest = surfaceContainerHighestDark,
+)
+
+private val mediumContrastLightColorScheme = lightColorScheme(
+    primary = primaryLightMediumContrast,
+    onPrimary = onPrimaryLightMediumContrast,
+    primaryContainer = primaryContainerLightMediumContrast,
+    onPrimaryContainer = onPrimaryContainerLightMediumContrast,
+    secondary = secondaryLightMediumContrast,
+    onSecondary = onSecondaryLightMediumContrast,
+    secondaryContainer = secondaryContainerLightMediumContrast,
+    onSecondaryContainer = onSecondaryContainerLightMediumContrast,
+    tertiary = tertiaryLightMediumContrast,
+    onTertiary = onTertiaryLightMediumContrast,
+    tertiaryContainer = tertiaryContainerLightMediumContrast,
+    onTertiaryContainer = onTertiaryContainerLightMediumContrast,
+    error = errorLightMediumContrast,
+    onError = onErrorLightMediumContrast,
+    errorContainer = errorContainerLightMediumContrast,
+    onErrorContainer = onErrorContainerLightMediumContrast,
+    background = backgroundLightMediumContrast,
+    onBackground = onBackgroundLightMediumContrast,
+    surface = surfaceLightMediumContrast,
+    onSurface = onSurfaceLightMediumContrast,
+    surfaceVariant = surfaceVariantLightMediumContrast,
+    onSurfaceVariant = onSurfaceVariantLightMediumContrast,
+    outline = outlineLightMediumContrast,
+    outlineVariant = outlineVariantLightMediumContrast,
+    scrim = scrimLightMediumContrast,
+    inverseSurface = inverseSurfaceLightMediumContrast,
+    inverseOnSurface = inverseOnSurfaceLightMediumContrast,
+    inversePrimary = inversePrimaryLightMediumContrast,
+    surfaceDim = surfaceDimLightMediumContrast,
+    surfaceBright = surfaceBrightLightMediumContrast,
+    surfaceContainerLowest = surfaceContainerLowestLightMediumContrast,
+    surfaceContainerLow = surfaceContainerLowLightMediumContrast,
+    surfaceContainer = surfaceContainerLightMediumContrast,
+    surfaceContainerHigh = surfaceContainerHighLightMediumContrast,
+    surfaceContainerHighest = surfaceContainerHighestLightMediumContrast,
+)
+
+private val highContrastLightColorScheme = lightColorScheme(
+    primary = primaryLightHighContrast,
+    onPrimary = onPrimaryLightHighContrast,
+    primaryContainer = primaryContainerLightHighContrast,
+    onPrimaryContainer = onPrimaryContainerLightHighContrast,
+    secondary = secondaryLightHighContrast,
+    onSecondary = onSecondaryLightHighContrast,
+    secondaryContainer = secondaryContainerLightHighContrast,
+    onSecondaryContainer = onSecondaryContainerLightHighContrast,
+    tertiary = tertiaryLightHighContrast,
+    onTertiary = onTertiaryLightHighContrast,
+    tertiaryContainer = tertiaryContainerLightHighContrast,
+    onTertiaryContainer = onTertiaryContainerLightHighContrast,
+    error = errorLightHighContrast,
+    onError = onErrorLightHighContrast,
+    errorContainer = errorContainerLightHighContrast,
+    onErrorContainer = onErrorContainerLightHighContrast,
+    background = backgroundLightHighContrast,
+    onBackground = onBackgroundLightHighContrast,
+    surface = surfaceLightHighContrast,
+    onSurface = onSurfaceLightHighContrast,
+    surfaceVariant = surfaceVariantLightHighContrast,
+    onSurfaceVariant = onSurfaceVariantLightHighContrast,
+    outline = outlineLightHighContrast,
+    outlineVariant = outlineVariantLightHighContrast,
+    scrim = scrimLightHighContrast,
+    inverseSurface = inverseSurfaceLightHighContrast,
+    inverseOnSurface = inverseOnSurfaceLightHighContrast,
+    inversePrimary = inversePrimaryLightHighContrast,
+    surfaceDim = surfaceDimLightHighContrast,
+    surfaceBright = surfaceBrightLightHighContrast,
+    surfaceContainerLowest = surfaceContainerLowestLightHighContrast,
+    surfaceContainerLow = surfaceContainerLowLightHighContrast,
+    surfaceContainer = surfaceContainerLightHighContrast,
+    surfaceContainerHigh = surfaceContainerHighLightHighContrast,
+    surfaceContainerHighest = surfaceContainerHighestLightHighContrast,
+)
+
+private val mediumContrastDarkColorScheme = darkColorScheme(
+    primary = primaryDarkMediumContrast,
+    onPrimary = onPrimaryDarkMediumContrast,
+    primaryContainer = primaryContainerDarkMediumContrast,
+    onPrimaryContainer = onPrimaryContainerDarkMediumContrast,
+    secondary = secondaryDarkMediumContrast,
+    onSecondary = onSecondaryDarkMediumContrast,
+    secondaryContainer = secondaryContainerDarkMediumContrast,
+    onSecondaryContainer = onSecondaryContainerDarkMediumContrast,
+    tertiary = tertiaryDarkMediumContrast,
+    onTertiary = onTertiaryDarkMediumContrast,
+    tertiaryContainer = tertiaryContainerDarkMediumContrast,
+    onTertiaryContainer = onTertiaryContainerDarkMediumContrast,
+    error = errorDarkMediumContrast,
+    onError = onErrorDarkMediumContrast,
+    errorContainer = errorContainerDarkMediumContrast,
+    onErrorContainer = onErrorContainerDarkMediumContrast,
+    background = backgroundDarkMediumContrast,
+    onBackground = onBackgroundDarkMediumContrast,
+    surface = surfaceDarkMediumContrast,
+    onSurface = onSurfaceDarkMediumContrast,
+    surfaceVariant = surfaceVariantDarkMediumContrast,
+    onSurfaceVariant = onSurfaceVariantDarkMediumContrast,
+    outline = outlineDarkMediumContrast,
+    outlineVariant = outlineVariantDarkMediumContrast,
+    scrim = scrimDarkMediumContrast,
+    inverseSurface = inverseSurfaceDarkMediumContrast,
+    inverseOnSurface = inverseOnSurfaceDarkMediumContrast,
+    inversePrimary = inversePrimaryDarkMediumContrast,
+    surfaceDim = surfaceDimDarkMediumContrast,
+    surfaceBright = surfaceBrightDarkMediumContrast,
+    surfaceContainerLowest = surfaceContainerLowestDarkMediumContrast,
+    surfaceContainerLow = surfaceContainerLowDarkMediumContrast,
+    surfaceContainer = surfaceContainerDarkMediumContrast,
+    surfaceContainerHigh = surfaceContainerHighDarkMediumContrast,
+    surfaceContainerHighest = surfaceContainerHighestDarkMediumContrast,
+)
+
+private val highContrastDarkColorScheme = darkColorScheme(
+    primary = primaryDarkHighContrast,
+    onPrimary = onPrimaryDarkHighContrast,
+    primaryContainer = primaryContainerDarkHighContrast,
+    onPrimaryContainer = onPrimaryContainerDarkHighContrast,
+    secondary = secondaryDarkHighContrast,
+    onSecondary = onSecondaryDarkHighContrast,
+    secondaryContainer = secondaryContainerDarkHighContrast,
+    onSecondaryContainer = onSecondaryContainerDarkHighContrast,
+    tertiary = tertiaryDarkHighContrast,
+    onTertiary = onTertiaryDarkHighContrast,
+    tertiaryContainer = tertiaryContainerDarkHighContrast,
+    onTertiaryContainer = onTertiaryContainerDarkHighContrast,
+    error = errorDarkHighContrast,
+    onError = onErrorDarkHighContrast,
+    errorContainer = errorContainerDarkHighContrast,
+    onErrorContainer = onErrorContainerDarkHighContrast,
+    background = backgroundDarkHighContrast,
+    onBackground = onBackgroundDarkHighContrast,
+    surface = surfaceDarkHighContrast,
+    onSurface = onSurfaceDarkHighContrast,
+    surfaceVariant = surfaceVariantDarkHighContrast,
+    onSurfaceVariant = onSurfaceVariantDarkHighContrast,
+    outline = outlineDarkHighContrast,
+    outlineVariant = outlineVariantDarkHighContrast,
+    scrim = scrimDarkHighContrast,
+    inverseSurface = inverseSurfaceDarkHighContrast,
+    inverseOnSurface = inverseOnSurfaceDarkHighContrast,
+    inversePrimary = inversePrimaryDarkHighContrast,
+    surfaceDim = surfaceDimDarkHighContrast,
+    surfaceBright = surfaceBrightDarkHighContrast,
+    surfaceContainerLowest = surfaceContainerLowestDarkHighContrast,
+    surfaceContainerLow = surfaceContainerLowDarkHighContrast,
+    surfaceContainer = surfaceContainerDarkHighContrast,
+    surfaceContainerHigh = surfaceContainerHighDarkHighContrast,
+    surfaceContainerHighest = surfaceContainerHighestDarkHighContrast,
+)
